// packages/registry/showcase/before/lib/upload-file-input/types.ts
export type FileFormatIconColor =
  | 'red'
  | 'orange'
  | 'yellow'
  | 'green'
  | 'teal'
  | 'blue'
  | 'purple'
  | 'pink'
  | 'gray';